# SavePoint/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SavePoint API",
    "version": "1.0.0",
    "description": "API REST do SavePoint: catálogo de jogos, comparação de preços entre lojas, histórico de preços, avaliações de usuários e lista de desejos."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/ofertas/{id}/historico": {
      "get": {
        "tags": [
          "Histórico de preços"
        ],
        "summary": "Listar histórico de preços da oferta",
        "operationId": "historico_buscarPorOfertaId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico do mais recente para o mais antigo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Historico"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Oferta não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Histórico de preços"
        ],
        "summary": "Registrar alteração de preço da oferta",
        "operationId": "historico_registrarAlteracaoPreco",
        "description": "Registra o novo preço no histórico, atualiza o preço atual da oferta e recalcula o desconto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Historico"
              },
              "example": {
                "preco": 79.9
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Historico"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Oferta não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
    "/usuarios/{usuarioId}/lista-desejos/{listaId}": {
      "delete": {
        "tags": [
          "Lista de desejos"
        ],
        "summary": "Excluir lista de desejos",
        "operationId": "listaDesejos_deletarLista",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "listaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}/lista-desejos/itens": {
      "post": {
        "tags": [
          "Itens da lista de desejos"
        ],
        "summary": "Adicionar jogo à lista de desejos",
        "operationId": "itemListaDesejos_adicionarItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemListaDesejos"
              },
              "example": {
                "precoAlerta": 79.9,
                "notificarOferta": true,
                "jogo": {
                  "id": 1
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListaDesejos"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou jogo já está na lista",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Lista ou jogo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}/lista-desejos/itens/{itemId}": {
      "get": {
        "tags": [
          "Itens da lista de desejos"
        ],
        "summary": "Buscar item da lista de desejos",
        "operationId": "itemListaDesejos_buscarItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListaDesejos"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Itens da lista de desejos"
        ],
        "summary": "Atualizar item da lista de desejos",
        "operationId": "itemListaDesejos_atualizarItem",
        "description": "Atualiza o preço de alerta e a notificação de oferta.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemListaDesejos"
              },
              "example": {
                "precoAlerta": 79.9,
                "notificarOferta": true,
                "jogo": {
                  "id": 1
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListaDesejos"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Jogo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "dataLancamento": {
            "type": "string",
            "format": "date"
          },
          "imagemCapa": {
            "type": "string"
          },
          "classificacaoIndicativa": {
            "type": "string"
          },
          "desenvolvedora": {
            "type": "string"
          },
          "genero": {
            "type": "string"
          },
          "requisitosMinimos": {
            "$ref": "#/components/schemas/RequisitosMinimos"
          },
          "resumoAvaliacao": {
            "$ref": "#/components/schemas/ResumoAvaliacao"
          },
          "ofertas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaJogo"
            }
          }
        }
      },
      "RequisitosMinimos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processador": {
            "type": "string"
          },
          "memoria": {
            "type": "string"
          },
          "placaDeVideo": {
            "type": "string"
          },
          "sistemaOperacional": {
            "type": "string"
          }
        }
      },
      "ResumoAvaliacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "notaMedia": {
            "type": "number",
            "format": "double"
          },
          "totalAvaliacoes": {
            "type": "integer",
            "format": "int32"
          },
          "percentualRecomendacao": {
            "type": "number",
            "format": "double"
          },
          "principaisElogios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "principaisCriticas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "resumoGeradoIA": {
            "type": "string"
          },
          "dataGeracao": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Loja": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "urlLoja": {
            "type": "string"
          },
          "urlLogo": {
            "type": "string"
          }
        }
      },
      "OfertaJogo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "precoOriginal": {
            "type": "number",
            "format": "double"
          },
          "precoAtual": {
            "type": "number",
            "format": "double"
          },
          "percentualDesconto": {
            "type": "number",
            "format": "float"
          },
          "loja": {
            "$ref": "#/components/schemas/Loja"
          },
          "historico": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Historico"
            }
          }
        }
      },
      "Historico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "preco": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ItemListaDesejos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "precoAlerta": {
            "type": "number",
            "format": "double"
          },
          "notificarOferta": {
            "type": "boolean"
          },
          "jogo": {
            "$ref": "#/components/schemas/Jogo"
          }
        }
      },
...
      "ErroResposta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "erro": {
            "type": "string"
          },
          "mensagem": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}